Switches from `npm ci` to `npm install`
Replaces `npm ci` with `npm install` in the workflow.

The `npm ci` command is designed for clean installations from a `package-lock.json` or `npm-shrinkwrap.json` file, and is typically used in automated environments to ensure reproducible builds. However, in certain scenarios, `npm install` might be more appropriate or flexible. This change aims to address potential issues or limitations encountered with `npm ci` in the current workflow setup, allowing for a broader range of dependency resolution and installation scenarios.